% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basis.p.r
\name{basis.p}
\alias{basis.p}
\title{Basic information for a vector}
\usage{
basis.p(a)
}
\arguments{
\item{a}{vector.}
}
\value{
a dataframe.
}
\description{
Show mean, sd, min, p25, p50, p75 and max for a vector.
}
\examples{
basis.p(1:5)
}
