% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z.test.r
\name{z.test}
\alias{z.test}
\title{Z test}
\usage{
z.test(b1, b2, se1, se2)
}
\arguments{
\item{b1}{coefficients in model 1.}

\item{b2}{coefficients in model 2.}

\item{se1}{standard error in model 1.}

\item{se2}{standard error in model 2.}
}
\value{
Z value and P value.
}
\description{
Test the difference between 2 coefficients.
}
\examples{
b1 <- 0.122275
b2 <- 0.358742
se1 <- (0.61492-(-0.36804))/3.92
se2 <- (0.883344-(-0.16313))/3.92
z.test(b1 = b1, b2 = b2, se1 = se1, se2 = se2)
# $z
# [1] -0.6456306
#
# $p
# [1] 0.5185186
}
